+++ src/main/java/pages/HomePage.java
@@ -1,4 +1,4 @@
-package page;
+package pages;
 
 import com.codeborne.selenide.SelenideElement;
